Refactor authentication and project handling and move functions to separate modules
