+++ src/auth.rs
@@ -1,3 +1,4 @@
+use crate::config::get_config_dir;
 use anyhow::{anyhow, Result};
 use colored::Colorize;
 use dialoguer::{Input, Password};
@@ -10,17 +11,6 @@ use std::path::PathBuf;
 pub struct Credentials {
     pub organization: String,
     pub pat: String,
-}
-
-fn get_config_dir() -> Result<PathBuf> {
-    let home_dir = dirs::home_dir().ok_or_else(|| anyhow!("Could not find home directory"))?;
-    let config_dir = home_dir.join(".azdocli");
-
-    if !config_dir.exists() {
-        fs::create_dir_all(&config_dir)?;
-    }
-
-    Ok(config_dir)
 }
 
 fn save_organization(organization: &str) -> Result<()> {
@@ -71,7 +61,7 @@ fn save_pat(pat: &str) -> Result<()> {
     let creds_path = get_credentials_file_path()?;
     let credentials_json = serde_json::to_string_pretty(&credentials)?;
 
-    let mut file = std::fs::File::create(&creds_path)?;
+    let mut file = fs::File::create(&creds_path)?;
     file.write_all(credentials_json.as_bytes())?;
 
     // Set restrictive permissions on the file
@@ -91,7 +81,7 @@ fn get_pat() -> Result<String> {
 
     if !creds_path.exists() {
         return Err(anyhow!(
-            "Not logged in. Please login first with 'azdocli login'"
+            "Not logged in. Please login first with 'ado login'"
         ));
     }
 
@@ -166,45 +156,4 @@ pub fn get_credentials() -> Result<Credentials> {
     let pat = get_pat()?;
 
     Ok(Credentials { organization, pat })
-}
-
-pub fn save_default_project(project: &str) -> Result<()> {
-    let config_dir = get_config_dir()?;
-    let config_file = config_dir.join("config.json");
-    let mut config = if config_file.exists() {
-        let config_content = fs::read_to_string(&config_file)?;
-        serde_json::from_str::<serde_json::Value>(&config_content)?
-    } else {
-        serde_json::json!({})
-    };
-
-    config["default_project"] = serde_json::Value::String(project.to_string());
-
-    fs::write(config_file, serde_json::to_string_pretty(&config)?)?;
-
-    Ok(())
-}
-
-pub fn get_default_project() -> Result<String> {
-    let config_dir = get_config_dir()?;
-    let config_file = config_dir.join("config.json");
-
-    if !config_file.exists() {
-        return Err(anyhow!("No default project configured. Please set one using 'azdocli project <project_name>' or use the --project argument"));
-    }
-
-    let config_content = fs::read_to_string(config_file)?;
-    let config: serde_json::Value = serde_json::from_str(&config_content)?;
-
-    match config["default_project"].as_str() {
-        Some(project) => Ok(project.to_string()),
-        None => Err(anyhow!("No default project configured. Please set one using 'azdocli project <project_name>' or use the --project argument"))
-    }
-}
-
-pub fn get_project_or_default(project_arg: Option<&str>) -> Result<String> {
-    match project_arg {
-        Some(project) => Ok(project.to_string()),
-        None => get_default_project(),
-    }
 }
